data-collection/pyproject.toml
[project]
name = "data-collection"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["aiohttp>=3.12.15", "python-dotenv>=1.1.1", "tenacity>=9.1.2"]
